requirements.txt
asgiref==3.7.2
Django==4.2.4
django-environ==0.10.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
mysqlclient==2.2.0
PyJWT==2.8.0
pytz==2023.3
sqlparse==0.4.4
